Implemented load/save of view plugins
Closes #164

+++ api/rust/prodbg/src/io.rs
@@ -20,7 +20,7 @@ pub struct CPDSaveState {
 pub struct CPDLoadState {
     pub priv_data: *mut c_void,
     pub read_int: fn(priv_data: *mut c_void, dest: *mut i64) -> LoadState,
-    pub read_double: fn(priv_data: *mut c_void, dest: *mut f32) -> LoadState,
+    pub read_double: fn(priv_data: *mut c_void, dest: *mut f64) -> LoadState,
     pub read_string: fn(priv_data: *mut c_void, dest: *mut c_char, max_len: i32) -> LoadState,
 }
